# perf: Add a zero-allocation version of the API