--- a/test/Moniker.PerformanceTests/NameGeneratorBenchmark.cs
+++ b/test/Moniker.PerformanceTests/NameGeneratorBenchmark.cs
@@ -9,4 +9,10 @@ public class NameGeneratorBenchmark
 
     [Benchmark]
     public string GenerateMoniker() => NameGenerator.Generate(MonikerStyle.Moniker);
+
+    [Benchmark]
+    public void GenerateMobyPair() => NameGenerator.Generate(MonikerStyle.Moby, out _, out _);
+
+    [Benchmark]
+    public void GenerateMonikerPair() => NameGenerator.Generate(MonikerStyle.Moniker, out _, out _);
 }