This is a follow-up to PR #8 

> > Incidentally, do you care to also add this? This way, the allocation of string becomes the entire responsibility of the caller, but it will require a modest expansion to the public API. Let me know if you're interested and I can propose something for prosperity, when I am bored next.
>
> I'm interested.

This PR renames `Utf8String` to `Chars`, makes it (mostly) public, and adds the following new zero-allocation APIs:

```c#
public static void Generate(MonikerStyle monikerStyle, out Chars adjective, out Chars noun) { /* … */ }
public static void GenerateMoniker(out Chars adjective, out Chars noun) { /* … */ }
public static void GenerateMoby(out Chars adjective, out Chars noun) { /* … */ }
```

Unfortunately, since [generic value tuples](https://learn.microsoft.com/en-us/dotnet/api/system.valuetuple-2?view=net-8.0) cannot carry `Chars`, one cannot simply return `(Chars, Chars)` or `(Chars Adjective, Chars Noun)` so out parameters are used instead. The other way to solve this would be to introduce a `CharsPair` type, but that would just add more code to maintain and test, and bring marginal benefit (though it can always be done later too).

`Utf8String` is renamed to `Chars` to encapsulate how data is stored internally. It has methods that permit writing into `char` and UTF-8 `byte` buffers for those who care to avoid an allocation; for those who don't care can simply call `ToString()`.

The benchmarks are updated and show roughly a 5x performance increase and _zero_ allocations with the new members:

| Method              | Mean     | Error    | StdDev   | Median   | Gen0   | Allocated |
|-------------------- |---------:|---------:|---------:|---------:|-------:|----------:|
| GenerateMoby        | 58.53 ns | 0.783 ns | 0.694 ns | 58.51 ns | 0.0043 |      55 B |
| GenerateMoniker     | 59.96 ns | 1.762 ns | 5.166 ns | 57.31 ns | 0.0042 |      53 B |
| GenerateMobyPair    | 11.84 ns | 0.221 ns | 0.246 ns | 11.85 ns |      - |         - |
| GenerateMonikerPair | 12.44 ns | 0.271 ns | 0.535 ns | 12.22 ns |      - |         - |

---
- BenchmarkDotNet v0.14.0, Windows 11 (10.0.22631.4037/23H2/2023Update/SunValley3)
- 13th Gen Intel Core i7-1370P, 1 CPU, 20 logical and 14 physical cores
- .NET SDK 8.0.400
  - [Host]     : .NET 8.0.8 (8.0.824.36612), X64 RyuJIT AVX2
  - DefaultJob : .NET 8.0.8 (8.0.824.36612), X64 RyuJIT AVX2